agent/requirements.txt
livekit-agents[silero,turn_detector,aws,azure,google]>=1.2.14